+++ src/Api/Core/FlixTv.Api.Application/Features/FavouriteMovies/Commands/DeleteFavouriteMovie/DeleteFavouriteMovieCommandHandler.cs
@@ -2,6 +2,7 @@
 using FlixTv.Api.Application.Interfaces.UnitOfWorks;
 using FlixTv.Api.Domain.Concretes;
 using MediatR;
+using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,12 +14,12 @@ namespace FlixTv.Api.Application.Features.FavouriteMovies.Commands.DeleteFavouri
     public class DeleteFavouriteMovieCommandHandler : IRequestHandler<DeleteFavouriteMovieCommandRequest, Unit>
     {
         private readonly IUnitOfWork unitOfWork;
-        private readonly IMapper mapper;
+        private readonly int userId;
 
-        public DeleteFavouriteMovieCommandHandler(IUnitOfWork unitOfWork, IMapper mapper)
+        public DeleteFavouriteMovieCommandHandler(IUnitOfWork unitOfWork, IHttpContextAccessor httpContextAccessor)
         {
             this.unitOfWork = unitOfWork;
-            this.mapper = mapper;
+            this.userId = Convert.ToInt32(httpContextAccessor.HttpContext?.User?.FindFirst("id")?.Value ?? "0");
         }
 
         public async Task<Unit> Handle(DeleteFavouriteMovieCommandRequest request, CancellationToken cancellationToken)
@@ -27,6 +28,9 @@ namespace FlixTv.Api.Application.Features.FavouriteMovies.Commands.DeleteFavouri
 
             if (favouriteMovie is null)
                 throw new Exception("The favourite movie was not found.");
+
+            if (favouriteMovie.UserId != userId)
+                throw new UnauthorizedAccessException("You can only delete your own favourite movie.");
 
             await unitOfWork.GetWriteRepository<UserMovieCatalog>().DeleteAsync(favouriteMovie);
             await unitOfWork.SaveAsync();
