+++ src/Api/Core/FlixTv.Api.Application/Features/Reviews/Commands/DeleteReview/DeleteReviewCommandHandler.cs
@@ -2,6 +2,8 @@
 using FlixTv.Api.Application.Interfaces.UnitOfWorks;
 using FlixTv.Api.Domain.Concretes;
 using MediatR;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,12 +15,14 @@ namespace FlixTv.Api.Application.Features.Reviews.Commands.DeleteReview
     public class DeleteReviewCommandHandler : IRequestHandler<DeleteReviewCommandRequest, Unit>
     {
         private readonly IUnitOfWork unitOfWork;
-        private readonly IMapper mapper;
+        private readonly UserManager<User> userManager;
+        private readonly int userId;
 
-        public DeleteReviewCommandHandler(IUnitOfWork unitOfWork, IMapper mapper)
+        public DeleteReviewCommandHandler(IUnitOfWork unitOfWork, IHttpContextAccessor httpContextAccessor, UserManager<User> userManager)
         {
             this.unitOfWork = unitOfWork;
-            this.mapper = mapper;
+            this.userManager = userManager;
+            this.userId = Convert.ToInt32(httpContextAccessor.HttpContext?.User?.FindFirst("id")?.Value ?? "0");
         }
 
         public async Task<Unit> Handle(DeleteReviewCommandRequest request, CancellationToken cancellationToken)
@@ -27,6 +31,13 @@ namespace FlixTv.Api.Application.Features.Reviews.Commands.DeleteReview
 
             if (review is null)
                 throw new Exception("Review was not found.");
+
+            var user = await userManager.FindByIdAsync(userId.ToString());
+
+            var roles = await userManager.GetRolesAsync(user!);
+
+            if (!roles.Contains("Admin") && !roles.Contains("Moderator") && userId != review.AuthorId)
+                throw new UnauthorizedAccessException("You are not authorized to delete this review.");
 
             await unitOfWork.GetWriteRepository<Review>().DeleteAsync(review);
             await unitOfWork.SaveAsync();
